+++ VeriTask/VeriTask/DataStorage.h
@@ -28,9 +28,9 @@ public:
 //	void printTaskToDisplay(); 
 	void fillData(vector<Task>);
 	void saveData(Task taskToSave);
-	void deleteData(double taskID);
+	void deleteData(unsigned long long int taskID);
 	bool searchData(string keyword);
-	void markData(double taskID, string status);
+	void markData(unsigned long long int taskID, string status);
 	void setTaskIndex(int taskIndex);
 	int getTaskIndex();
 };
